% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stimuli.R
\name{create_stimuli}
\alias{create_stimuli}
\title{Create stimuli values.}
\usage{
create_stimuli(stimuli_constant = 50, stimuli_max = 90, num_stimuli = 9)
}
\arguments{
\item{stimuli_constant}{The stimuli value that will remain constant.}

\item{stimuli_max}{The maximum stimuli value.}

\item{num_stimuli}{Total number of stimuli. If an even number is used, the
number of stimuli increases by one.}
}
\value{
A tibble
}
\description{
\code{create_stimuli()} creates a dataset using the method of constant stimuli.
This approach generates varying values by using equally spaced ratios between
the constant stimuli and the stimuli with the maximum value. The ratios are
then replicated with the constant stimuli as the maximum value to create
the stimuli that are smaller than the constant stimuli.
}
\examples{
create_stimuli()

}
