% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_heatmap_data.R
\name{generate_heatmap_data}
\alias{generate_heatmap_data}
\title{Function to create heatmap with given stimuli values}
\usage{
generate_heatmap_data(
  x = 1:10,
  y = 1:10,
  c = 0.3,
  f,
  stimuli,
  num_iters = 20,
  ...
)
}
\arguments{
\item{x, y}{A vector of grid coordinates for the heatmap.}

\item{c}{A mixture parameter for the proportion of uniform randomness.}

\item{f}{A function for the true shape over the heatmap. Values will be
scaled between 0 and 100.}

\item{stimuli}{A tibble generated by \code{create_stimuli()}.}

\item{num_iters}{Number of iterations to refine stimuli coordinates.}

\item{...}{Additional arguments that are supplied to \code{f}.}
}
\value{
A tibble with (X,Y) coordinates, z, and stimuli pair identifiers.
}
\description{
\code{generate_heatmap_data()} creates a data set that reasonably inserts known
stimuli value pairs into a chosen function. The function will also try to
balance the locations of the stimuli values across the grid coordinates.
}
\examples{
generate_heatmap_data(f=\(x,y) sqrt(36-(x-mean(x))^2-(y-mean(y))^2))
}
